requirements.txt
transformers
torch
datasets
tqdm
python-dotenv
pyyaml
accelerate
sentencepiece
huggingface_hub
